Refactoring and DOC
